requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
qrcode==8.2
sqlparse==0.5.5
whitenoise==6.11.0
